rendu

+++ libft/srcs/put/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: alevasse <alevasse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/11 10:38:43 by alevasse          #+#    #+#             */
-/*   Updated: 2022/04/21 09:14:58 by alevasse         ###   ########.fr       */
+/*   Updated: 2022/05/16 13:53:01 by alevasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putnbr_base(unsigned int n, char *base)
 {
-	if (n >= ft_strlen(base))
+	if ((int)n >= ft_strlen(base))
 		ft_putnbr_base(n / ft_strlen(base), base);
 	ft_putchar(base[n % ft_strlen(base)]);
 }
